refactor(converter): `to_speckle` refactor!!
