--- bpy_speckle/convert/to_speckle/__init__.py
@@ -1,3 +1,0 @@
-from .mesh import export_mesh
-from .curve import export_curve, export_ngons_as_polylines
-from .empty import export_empty
