--- bpy_speckle/convert/to_speckle/default.py
@@ -1,5 +1,0 @@
-import bpy, bmesh, struct
-
-
-def export_default(blender_object, scale=1.0):
-    return None
