--- bpy_speckle/convert/to_speckle/empty.py
@@ -1,5 +1,0 @@
-import bpy, bmesh, struct
-
-
-def export_empty(blender_object, data, scale=1.0):
-    return None
